VisionCore: Minor improvements.
